pyproject.toml
[project]
name = "autoria"
version = "0.1.0"
description = ""
authors = [
    {name = "iriadda",email = "iriadda@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "django-filter (>=25.1,<26.0)",
    "better-profanity (>=0.7.0,<0.8.0)",
    "simplejwt (>=2.0.1,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "djangochannelsrestframework (>=1.3.0,<2.0.0)",
    "daphne (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",



]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

